Implement proper fork/exec with argument passing and add ls command (#81)
* Added a very basic (non-functional) ls

* Minor cleanup

* Tested -O0 on Kernel, and fixed one minor bug

* Added ls in .gitignore

* Removed mode_t header and moved it to types.h

* Added open as user syscall

* Added some pushes and pops

* Moved errno.h to uapi

* Better error messaging on execve error

* updated ls a bit more to work on the userspace

* Added a comment to libc Makefile

* Got a working /bin/ls

* Argv & argc are now functional and can be passed along

+++ userspace/lib/libc/include/libc/string.h
@@ -6,7 +6,7 @@
 #define memcpy(dest, src, n) __builtin_memcpy((dest), (src), (n))
 #define memset(s, c, n) __builtin_memset((s), (c), (n))
 
-int strlen(char const*);
+size_t strlen(char const*);
 
 int strcmp(char const*, char const*);
